main: Add apps list view
The list consists of a RecyclerView which is populated with data from Parse. We had to create a custom adapter to populate the RecyclerView. Also, we have added a FloatingActionButton. It will be used to add new data to the system.

This view is contained in a fragment, since we will add a different layout for large screens, which will contain two fragments.

Signed-off-by: Álvaro Orduna León <alvaro.orduna.92@gmail.com>